extract generic RewilderModal from ConnectWalletModal
