+++ src/components/ConnectWalletModal.js
@@ -1,20 +1,11 @@
 import { useEthers } from "@usedapp/core";
-import { useEffect, useCallback } from "react";
+import { useEffect } from "react";
 import { walletconnect } from "../lib/connectors";
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faTimes } from '@fortawesome/free-solid-svg-icons';
 
+import RewilderModal from './RewilderModal';
 
 function ConnectWalletModal({ onOpen, isOpen, onClose }) {
   const { activate, activateBrowserWallet, error } = useEthers();
-
-  useEffect(() => {
-    if (isOpen) {
-      document.body.classList.add('overlay');
-    } else {
-      document.body.classList.remove('overlay');
-    }
-  }, [isOpen]);
 
   useEffect(() => {
     if (error) {
@@ -43,46 +34,23 @@ function ConnectWalletModal({ onOpen, isOpen, onClose }) {
     onClose();
   };
 
-  // ESC key closes
-  const escFunction = useCallback((event) => {
-    if(event.keyCode === 27) {
-      onClose();
-    }
-  }, []);
-
-  useEffect(() => {
-    document.addEventListener("keydown", escFunction, false);
-
-    return () => {
-      document.removeEventListener("keydown", escFunction, false);
-    };
-  }, []);
-
   return (
-    <div>
-    { 
-      <div className={"connect-wallet-popup"+(isOpen?" active":"")}>
-        <div className="connect-wallet-close" onClick={onClose}>
-          <FontAwesomeIcon icon={faTimes} />
-        </div>
-        <h4>Connect to a wallet</h4>
-        <div className="connect-option">
-          <a href="#" onClick={clickMetamask}>
-            <img src="/assets/img/icon/metamask-icon.svg" alt="Metamask" />
-            <span>Metamask</span>
-          </a>
-          <a href="#" onClick={clickWalletConnect}>
-            <img src="/assets/img/icon/wallet-connect-icon.svg" alt="WalletConnect" />
-            <span>WalletConnect</span>
-          </a>
-        </div>
-        <p>
-          By connecting a wallet, you agree to Rewilder’s 
-          <a href="#">Terms of Service</a> and <a href="#">Privacy Policy.</a>
-        </p>
+    <RewilderModal onOpen={onOpen} isOpen={isOpen} onClose={onClose} title={"Connect to a wallet"}>
+      <div className="connect-option">
+        <a href="#" onClick={clickMetamask}>
+          <img src="/assets/img/icon/metamask-icon.svg" alt="Metamask" />
+          <span>Metamask</span>
+        </a>
+        <a href="#" onClick={clickWalletConnect}>
+          <img src="/assets/img/icon/wallet-connect-icon.svg" alt="WalletConnect" />
+          <span>WalletConnect</span>
+        </a>
       </div>
-    }
-    </div>
+      <p>
+        By connecting a wallet, you agree to Rewilder’s 
+        <a href="#">Terms of Service</a> and <a href="#">Privacy Policy.</a>
+      </p>
+    </RewilderModal>
   );
 }
